+++ lib/s_extensions.dart
@@ -4,3 +4,5 @@ export 'extensions/number_ext.dart';
 export 'extensions_context.dart';
 export 'extensions/string_ext.dart';
 export 'extensions/datetime_ext.dart';
+export 'extensions/map_ext.dart';
+export 'extensions/list_ext.dart';